fix(node-full): Drop datasource import variable
The "Datasource" dropdown variable did not work at all.
Mostly because during import all references to that
variable were replaced by an import "input" variable.

So let's just drop the import input variable entirely (so
during import no request is made to select a datasource).
That way the dropdown works nicely again for everything.

Also rename it to lower case, because that seems to usual
convention for dropdown variables.